Unset $VIM and $VIMRUNTIME in the launched terminals
These are inherited from the invoking Vim instance, but the values cause problems if a different Vim version is started inside the terminal, because existing values override the built-in defaults. So if I start a Vim 7.4 from a terminal that was started from Vim 8.2, 'runtimepath' suddenly includes /usr/local/share/vim/vim82 instead of /usr/share/vim/vim74, and default plugins that do not belong to that version are loaded.
(I noticed the problem because I have a workaround for the matchit.vim plugin in older < 8.1 Vims that removes select-mode mappings, and sometimes a launched Vim 7.4 would complain that these mapings do not exist, and it turned out that that Vim instance would use the fixed 8.2 pack plugin through the inherited $VIMRUNTIME.)