Set b:cwd to indicate the root directory the terminal was opened in
Vim doesn't seem to have a notion of current working directory for terminal buffers; getcwd() will return the global one / the one of the last visited file buffer if 'autochdir' is set. To make the directory that the terminal was opened in available, set this buffer variable. (This doesn't mean that the terminal still uses that directory; it might have chdir'ed elsewhere.)